requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --extra dev -o requirements.txt
altair==6.0.0
    # via streamlit
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   google-genai
    #   httpx
    #   starlette
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==7.0.5
    # via streamlit
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via
    #   streamlit
    #   uvicorn
cryptography==47.0.0
    # via google-auth
distro==1.9.0
    # via google-genai
fastapi==0.135.2
    # via cle-engine (pyproject.toml)
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-auth==2.49.2
    # via google-genai
google-genai==1.73.1
    # via cle-engine (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   cle-engine (pyproject.toml)
    #   google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.18.1
    # via altair
numpy==2.4.4
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==26.0
    # via
    #   altair
    #   pytest
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==12.1.1
    # via streamlit
pluggy==1.6.0
    # via pytest
protobuf==6.33.6
    # via streamlit
psycopg2-binary==2.9.11
    # via cle-engine (pyproject.toml)
pyarrow==23.0.1
    # via streamlit
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   google-genai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.20.0
    # via pytest
pytest==9.0.3
    # via
    #   cle-engine (pyproject.toml)
    #   pytest-asyncio
pytest-asyncio==1.3.0
    # via cle-engine (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
pytz==2026.1.post1
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.33.0
    # via
    #   cle-engine (pyproject.toml)
    #   google-auth
    #   google-genai
    #   streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
ruff==0.15.8
    # via cle-engine (pyproject.toml)
six==1.17.0
    # via python-dateutil
smmap==5.0.3
    # via gitdb
sniffio==1.3.1
    # via google-genai
starlette==1.0.0
    # via fastapi
streamlit==1.55.0
    # via cle-engine (pyproject.toml)
tenacity==9.1.4
    # via
    #   google-genai
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.5
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   fastapi
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
uvicorn==0.42.0
    # via cle-engine (pyproject.toml)
websockets==16.0
    # via google-genai
